pyproject.toml
[project]
name = "github-scraping-ai"
version = "0.1.0"
description = "AI-powered GitHub project discovery"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
